SIS2: Rescale wind stresses on ice when part_size changes
  Added code that rescales the wind stresses on the ice and the open water when
the thermodynamics are not interspersed with the dynamics and transport, so
that the total momentum imparted to the ice-ocean system is conserved.  This
would change answers when INTERSPERSED_ICE_THERMO is false, but as this is
not yet true in any of the test cases, there are no answer changes.