SIS2: +*Use Ulm capabilities in restart files.
  Rewrote the code doing restarts of variables (like enthalpy and salinity) that
vary with the layers in the vertical, so that they are now output as 4-d arrays
rather than lists of 3-d arrays.  The variables like t_ice1, t_ice2, ... are no
longer written to restart files, although they might be read if the new
variables are not found.  Answers are not bitwise identical if any previous SIS2
restart files are used, but as I believe that no one is doing this yet, it
should not be a problem.  In addition, RESTARTFILE has been added as a run-time
variable with a default that reproduces the previous name.  Also
Energy_melt_enthS replaces e_to_melt_TS in the stock code for heat, which does
change answers in diagnostic output.  Despite these changes, the answers for
all test cases are unchanged, but there are new SIS_parameter_doc.all files.