Changed ice_ocean_SISn cases to avoid LEGACY_SPLIT
  Modified the MOM_input files for many of the ice_ocean_SIS and ice_ocean_SIS2
test cases so that none now use the LEGACY_SPLIT option.  These are now all
based closely on the ocean components of the MOM6i_1deg test cases.  This
changes the answers and ascii output files for all of these test cases, and new
reference solutions have been checked in.