MOM6: +Thermodynamic timesteps can now span coupled steps
  Added the ability for thermodynamic timesteps to cover multiple coupling
timesteps.  This is enabled at run-time by setting SPAN to be true and setting
DT_THERM.  All answers are bitwise identical by default, but the documentation
of this feature in the MOM_parameter_doc files causes them to change.  There are
several new interfaces related to these changes: adding forcing_accumulate and
splitting forcing_diagnostics into two subroutines, forcing_diagnostics and
mech_forcing_diags, both of which must be called. A new element was added to the
forcing type and another was renamed and changed form an integer to a logical
variable.  In addition, the time ranges for diagnostics of the bottom boundary
layer properties were corrected. By default these modifications do not change
solutions.