Added SLight MOM_override file for OM4_025
  Added a new file, MOM_override_SLight, demonstrating the changes that would be
needed to replace the Z* coordinate with the SLight coordinate (a streched
coordinate in light water over an isopycnal interior), including using the
options to revert to the streched coordinate in near-surface thermal inversions
and to set maximum interface depths.  This file has been tested, but is not used
with the current reference solutions.