Removed 8 test cases from routine testing.
  Because of GFDL's new and smaller computer system, we are removing 8 test
cases that are large or costly and do not often detect errors from our routine
test suite.  To do this, we have simply deleted their representative ocean.stats
and seaice.stats files from the repository so that they are not routinely
remade, although it is intended that all of these test cases will continue to
work.  We have decided to retain the MOM_parameter_doc and SIS_parameter_doc
files to document the parameters that were last used for these test cases,
although these may become out of date as new parameters are added or default
values are changed.