Added Davis strait and remove Canadian Archipelago
Added Davis strait and removed Canadian Archipelago. This modification
accords with recommendations from experts who noted that the original
CAA transport is not well constrained by the observations.  In contrast,
the Davis St. transport was recently detailed in a paper from 2014, based
on measurements from 2004-2010.  This new transport will be added
to the CMIP6 request, with the original Canadian Archipelago transport
removed.