MOM6: +Split MOM_coord_initialization from MOM_fixed_init
  Created a new module, MOM_coord_initialization, to handle the initialization
of any time-invariant aspects of the ocean model's vertical coordinate out of
the code that had been in MOM_fixed_initialization.F90.  As a part of this, a
spelling error was corrected in the documentation of one of the run-time
parameters, so new MOM_parameter_doc files were checked in for many experiments.
All answers are bitwise identical.
MOM6 commit NOAA-GFDL/MOM6@90580e45e04715e7a47bf752dd9648beec0f669b