SIS2: +Added WRITE_GEOM option for SIS2.
  Added the option for SIS2 to write out its geometry (masks and grid metrics)
file, similarly to MOM6.  Also disabled the test of the grid copy, which
eliminates some duplicate lines in the SIS_parameter_doc files.  When the
ice geometry file is written, whether it is decomposed or a single file is
determined by the parameter PARALLEL_RESTARTFILES, which is inhereted from
MOM6, although the description is a bit confusing in this case. All solutions
are identical, but a new optional run-time parameter was added.
SIS2 commit NOAA-GFDL/SIS2@e34a7335d84e3dc45835baee95a4f0876c3280ce