SIS2: +Renamed IST%coszen to IST%coszen_nextrad
  Renamed the element coszen in the ice state type to make it clear what time
period it refers to.  Also moved the code setting the ocean albedo, either
via the model time or from the coszen array, into a new routine,
set_ocean_albedo, which will soon be added as a call during initialization.
Also clarified the comments for the diagnostic coszen, which changes some of
the avialable_diag fields.  All answers are bitwise identical, although an interface was changed and the
name of an element in a shared type was changed.
SIS2 commit NOAA-GFDL/SIS2@870950214bc2bf83b772a89227ec2e2612470bd8