SIS2: +Added the runtime option ICEBERG_WINDSTRESS_BUG
  Added the runtime option ICEBERG_WINDSTRESS_BUG.  If true, the current
solutions, which pass an old set of ice-ocean stresses to the icebergs as though
they were current wind-stresses, are retained.  If this is set to false, the
icebergs are given the latest wind stresses on an A-grid.  By default, all
solutions are bitwise identical, but the SIS_parameter_doc.all files change.
SIS2 commit NOAA-GFDL/SIS2@8c3001039e7827e660241d15743bbdf51c672654