SIS2: +Added VERONA_COUPLER runtime argument
  Added a runtime argument indicating that the Verona or earlier version of
the coupler code is being used.  By default this does not change anything, and
if a newer version of the coupler is used, the answers do not change, even in
coupled models, but with Verona or older versions of the coupler, errors occur
in coupled models unless this VERONA_COUPLER=true.  This change modifies the
SIS_parameter_doc files.
SIS2 commit NOAA-GFDL/SIS2@6912f9029ffed597091b5df9d7500d91641bc6ab