SIS2: Added FIA%Tskin_avg & altered TS diagnostic
  Added a new array (Tskin_avg) to the fast_ice_avg type for the area-weighted
average of the skin temperature atop the ice or snow, or 0 C if there is no ice
coverage in a cell.  Because this is now a proper average of the skin
temperatures as given to the atmosphere, it is changed modestly from its
previous value.  In addition, the calls to save this diagnostic have been moved
from post_ice_state_diagnostics into post_flux_diagnostics.  Because the order
in which the diagnostics are registered have change, this leads to changes in
the SIS.available_diags files. This is only a diagnostic change; the solutions
themselves are bitwise identical.
SIS2 commit NOAA-GFDL/SIS2@e329db1007b77bb8f216b9a49c3bf6cd8f6d844a