+Updated code versions to warsaw & verona_201701
  Updated the version of the code that MOM6-examples points to the warsaw
versions of all components except for the coupler, which includes one bug fix
after warsaw and LM3/land_lad2 which is left at verona_201701.  MOM6, SIS2 and
icebergs are all following the latest versions of dev/master.  All answers in
the test cases are unchanged, but interfaces have evolved, notably the cpp flag
_USE_LEGACY_LAND_ must be defined to compile any cases with a land component.