+SIS2 uses symmetric memory in OM_360x320_C180
  Changed the land_ice_ocean_LM3_SIS2/OM_360x320_C180 test case to use symmetric
memory in SIS2.  This does not change answers in the short test run, but it will
change this configuration's ability to reproduce across restarts because it does
not have a land point along its southern boundary.  SIS_parameter_doc.layout has
been updated, and a copy of SIS2_memory.h added to this directory.