Added coupled_.../Concurrent_ice_1deg test case
  Added the new coupled_AM2_LM3_SIS2/Concurrent_ice_1deg test case, which uses
concurrent-ice coupling, in which the fast ice is updated with the atmosphere at
the same time the slow ice is updated with the ocean using lagged fluxes from
the previous fast ice update.  The reference solutions have not yet been added
because this test requires an updated version of the coupler, which should
become available with the first patch of the Warsaw release.