Updated coupler to add Concurrent_ice_1deg
  Updated the version of coupler to allow the addition of the
Concurrent_ice_1deg test case.  Be advised that this requires the use of the
SIS2_interface_rwh branch of SIS1 from gitlab to run the old CM2G63L test case,
and that the required changes to SIS1 were omitted from the warsaw release.
Moreover, because we can not link to gitlab code in the MOM6-examples checkout,
I do not know how to automatically correct this problem.  All existing answers
are bitwise identical.