Include wmo as part of refine diags
A requested variable 'wmo' the upwards mass transport is a requested
CMIP6 diagnostic which as of yet is not available from the prognostic
model in z* space. Transport across a vertical interface in hybrid
coordinates can be slightly different than what users may canonically
think of as a upward (towards the surface) direction which is orthogonal
to strictly meridional/zonal directions. To be consistent with this
colloquial understanding of vertical direction, wmo is calculated from
the convergence of umo and vmo which have been remapped to a z* coordinate.