Adds Baltic OM4_05 test case
This test case is intended for development and testing (not for scientific
use). The model domain centers on the Baltic Sea and the grid has been cut
out of the global OM4_05 test case and uses the same MOM_input and
SIS_input. The model is small enough to be run relatively quickly on a
workstation. A script is included to update various input fields if they
change for OM4_05.