MOM6: +Moved u, v, and h diagnostics & other changes
  Moved the diagnostics "u", "v", and "h" into MOM_diagnostics, writing them in
calculate_diagnostic_fields, and added new diagnostics "u_dyn", "v_dyn", and
"h_dyn", which are written at the end of the dynamics step.  The frequency with
which "u", "v", and "h" are written can decrease if there are multiple dynamics
time step per tracer step, which in turn can change time-averaged diagnostics.
All solutions are bitwise identical, but there are additional diagnostics and
the available_diags files change.

- NOAA-GFDL/MOM6@07eb6f1 +Moved u, v, and h diagnostics
- NOAA-GFDL/MOM6@a62a171 +Added symmetrize argument to set_viscous_ML
- NOAA-GFDL/MOM6@05d62a0 +Further refactoring of MOM_control_struct
- NOAA-GFDL/MOM6@1d1530c +Eliminated forces argument from bt_mass_source
- NOAA-GFDL/MOM6@3518f70 +Added symmetrize argument to set_viscous_BBL
- NOAA-GFDL/MOM6@15b8da4 Merge branch 'dev/gfdl' into restructure_step_MOM
- NOAA-GFDL/MOM6@50c41d4 +Added step_MOM_dynamics
- NOAA-GFDL/MOM6@8b7c043 +Obsoleted BT_MASS_SOURCE_LIMIT
- NOAA-GFDL/MOM6@dca91ee +Rearrange MOM_control_struct and MOM_state_type
- NOAA-GFDL/MOM6@6680003 Corrected tracer diagnostic metadata formatting