MOM6: +Restructured step_MOM(2)
  Completed the serious of changes to add the ability to do updates of the
dynamics and thermodynamics separately, including recombining the
ocean_state_type and MOM_control_struct to simplify the calls and interfaces.
This also includes adding the options that can be used to verify that for
coupled test cases, the solutions are identical regardless of whether the
dynamics and thermodyanmics are updated together or via separate calls to
step_MOM with the equivalent sequence order. The MOM_parameter_doc.all files are
changed by these commits, but no answers change.
- NOAA-GFDL/MOM6@783259b Merge branch 'dev/gfdl' into restructure_step_MOM
- NOAA-GFDL/MOM6@3701c51 Renamed cycle_time to time_in_cycle in MOM_CS
- NOAA-GFDL/MOM6@e16b96d +Added SINGLE_STEPPING_CALL coupled driver option
- NOAA-GFDL/MOM6@9afabcb Pass dt_forcing to some step_MOM calls
- NOAA-GFDL/MOM6@7b6ce15 +Added cycle_length optional argumetn to step_MOM
- NOAA-GFDL/MOM6@ed079a1 +Added MOM_surface_chksum
- NOAA-GFDL/MOM6@7d5f575 Correct Leith term calculation index ranges
- NOAA-GFDL/MOM6@ad3ad48 Merge pull request #717 from angus-g/seamount-fix
- NOAA-GFDL/MOM6@1df8e3b Restore missing constant in seamount initialization
- NOAA-GFDL/MOM6@133e5b5 Merge pull request #716 from adcroft/fix-sponge-diag-issue-714
- NOAA-GFDL/MOM6@478006f Fix uninitialized data in sponge diagnostic
- NOAA-GFDL/MOM6@b2ea2fb +Merged ocean_state_type and MOM_control_struct