MOM6: (*)Do not recalculate DTBT every coupled timestep
  Eliminated the requirement to recalculate DTBT every coupled timestep.  By
default, this still happens, but explicitly setting DTBT_RESET_PERIOD to be
longer than the coupled timestep now results in less frequent updates to DTBT.
Also modified the log_param documentation of DTBT_RESET_PERIOD to reflect this
new behavior.  This could change answers, but it just happens that the values of
DTBT are not changed with the existing suite of test cases.  The
MOM_parameter_doc files are altered by this change.
 - NOAA-GFDL/MOM6@a680bf7 (*)Do not recalculate DTBT every coupled timestep
 - NOAA-GFDL/MOM6@5579d26 +Allow DTBT recalculation interval to span steps
 - NOAA-GFDL/MOM6@1efd1f9 +Add optional arg salt to allocate_forcing_type
 - NOAA-GFDL/MOM6@5312ee1 Reordered initialize_ALE_sponge_fixed declarations
 - NOAA-GFDL/MOM6@17a28a6 Corrected openMP calls in vertvisc
 - NOAA-GFDL/MOM6@8e0debd Corrected openMP calls in MOM_KPP.F90
 - NOAA-GFDL/MOM6@e0a2328 (*+) Corrected a bug setting rigidity from icebergs
 - NOAA-GFDL/MOM6@2d07aa0 Removed unused module use statements from MOM.F90