Changed double_gyre DEPTH_LIST parameter settings
  Changed the REQUIRE_DEPTH_LIST_CHECKSUMS and UPDATE_DEPTH_LIST_CHECKSUMS
parameter settings for the double_gyre test case so that the correct depth_list
file will be created if an older one or one generated with a different compiler
is already at the indicated location.  All answers are bitwise identical, but
the MOM_parameter_doc files for the double_gyre test case are altered.