Fix calculation of Pacific Equatorial Undercurrent transport
Huan Gao and Rong Zhang noted that the transport in the equatorial
undercurrent should be around ~24 Sv in the annual mean, but in
the actual diagnostic after postprocessing, it was around -8Sv. This
was traced down to errors in the strait_transport_timeseries.

This commit fixes the logical and indexing errors in that script namely
in the way that the z-index to integrate to and the zeroing out of any
deeper layers. For the year 650 of the CM4 PI-Control experiment, the
routine now returns an annually averaged transport of 25 Sv.