Corrected SIS_input_files
  Corrected the SIS_input files to write both SIS_parameter_doc and
SIS_fast_paramater_doc files, and to explicitly set two parameters for the
Concurrent_ice test case.  Also added a non-default (but not logged) parameter
to the MOM_input file for the ISOMIP/rho test case.  All answers and output
files are back to where they should be, both with the current version of the
SIS2 code base, and with a version that changes some SIS2 default values.