Update ISOMIP input files
  Updated the MOM_input files for the ISOMIP/layer and ISOMIP/rho test cases and
added MOM_parameter_doc files for ISOMIP/layer.  Along with some fixes to the
MOM6 code in a pending PR, ISOMIP/layer now works.  The changes to MOM_input for
ISOMIP/rho are for consistency with the MOM_parameter_doc.short file for this
case.  The changes to the referenced input files do change answers for
ISOMIP/layer.