MOM6: +Post tracer diagnostics at the right point
  This MOM6 PR returns the call to post 3-d tracer concentration diagnostics to
the correct point in the MOM6 algorithm, and adds a new set of alternative
tracer concentration diagnostics just after the horizontal advection and
diffusion where the primary diagnostics had been before this PR.  Because of the
new tracer diagnostics, there are new entries in the available_diags files, and
the values reported with the primary tracer diagnostics, including temperature
and salinity, will have changed, but the model solutions themselves do not
change.  This PR addresses MOM6 issue #1206, which can now be closed.  The
commits in this PR include:

- NOAA-GFDL/MOM6@fd93dd72c Merge pull request #1207 from ashao/move_tracer_concentration_post
- NOAA-GFDL/MOM6@e45f73f4e Diagnostic for post-transport tracer concentration
- NOAA-GFDL/MOM6@334250f4f Merge branch 'move_tracer_concentration_post' of github.com:ashao/MOM6 into move_tracer_concentration_post
- NOAA-GFDL/MOM6@503eaa31d Update MOM_tracer_registry.F90
- NOAA-GFDL/MOM6@e93c031f1 Neglected subroutine in new declaration
- NOAA-GFDL/MOM6@0178309cc Merge branch 'dev/gfdl' of github.com:NOAA-GFDL/MOM6 into move_tracer_concentration_post
- NOAA-GFDL/MOM6@304d0c76e Post tracer concentration when synced