*Avoid USE_LEGACY_DIABATIC_DRIVER with global_ALE
  Modified the MOM_input file for the global_ALE test case to avoid using the
legacy diabatic driver with the global_ALE test cases, to expand the range of
options that are tested with the MOM6-examples regression suite.  The
MOM_parameter_doc files were updated for the global_ALE test cases, and the
MOM_parameter_doc files for infrequently run test cases are also updated.  The
answers in the global_ALE test cases are mathematically equivalent, but are
being deliberately changed at the level of roundoff.