Update MOM_parameter_files for less-tested cases
  Updated the MOM_parameter_doc files for 8 test cases in MOM6-examples that are
not routinely run in the gitlab pipelines, mostly to reflect a different
position for the parameter ICE_SHELF in these files and some new parameters
arising from recent commits.  Also updated the MOM_input files for the
ISOMIP/layer and ISOMIP/rho test cases to be more consistent with the order of
parameters in the MOM_parameter_doc files associated with the repositioning of
the code that initializes the ice shelves.  All answers are bitwise identical.