Update tides MOM_input and MOM_parameter_doc files
  Updated the MOM_Parameter_doc files in test cases that are in MOM6-examples
but are not routinely run as a part of the MOM6-examples regression tests, to
reflect recent updates to the MOM6 code, and modified the MOM_input file for
the tides_025 test case to shorten it.  All answers are bitwise identical in
the typical test cases, but one of the infrequently used tests is shorter.