SIS2: +Add landfast ice parameterization
  This series of commits to SIS2 adds the option to use a landfast ice
parameterization with C-grid versions of the SIS2 code, and it fixes one issue
with restarting with a particular initialization mode.  The new landfast ice
option is enabled with the new runtime parameter LEMIEUX_LANDFAST. By default
there are no changes to answers, but there are new entries in SIS_parameter_doc
files.  The version of SIS2 used by MOM6-examples and the SIS_parameter_doc.all
files were updated.  The commits in this PR include:

- NOAA-GFDL/SIS2@6d47c6c Merge pull request #136 from ESMG/dev/esmg
- NOAA-GFDL/SIS2@97992d0 Fix restart problem with latitude initialization.
- NOAA-GFDL/SIS2@057cf2c Added missing abs functions.
- NOAA-GFDL/SIS2@091eff1 Fix a comment
- NOAA-GFDL/SIS2@7be9038 Still fussing with landfast ice.
- NOAA-GFDL/SIS2@20e679b Fix an omission.
- NOAA-GFDL/SIS2@69d0011 Stab at adding the Lemieux landfast ice option.
- NOAA-GFDL/SIS2@b0200fd Fix that Liz needed for restarts.
- NOAA-GFDL/SIS2@d6fd875 Fix to tracer advection.
- NOAA-GFDL/SIS2@76dbcbb +Added TRACER_ADV_REFILL_MASSLESS option