Explicit module loading
A change to the Gaea module environment was corrupting the user's
environment in the MOM gitlab testing, causing all module commands to
fail and making it impossible to build and run the models.

This may be caused by the introduction of `readonly` variables in the
`/etc/bash.bashrc.local` file, which prevented critical modules from
being loaded (presumably to prevent double `module load` statements).
It is possible these were incorrectly inherited to noninteractive Gaea
sessions.

This patch introduces a more explicit module setup.  The process is
described below in more detail, in case further changes are needed.

1. module purge

   Clear out any existing modules (but not the readonly flags)

2. source /opt/cray/pe/modules/default/etc/modules.sh

   Bootstrap the module system and add and following modulepaths:

     - /opt/modulefiles: Compilers
     - /opt/cray/modulefiles: Lower level filesystem
     - /opt/cray/pe/modulefiles: PrgEnv-* and libraries

3. Load the following modules:

     - craype-network-aries
     - eproxy
     - PrgEnv-${compiler}
     - craype-broadwell
     - cray-mpich

4. Select compiler version and load the version-dependent modules:

     - cray-hdf5
     - cray-netcdf

Much of this process replicates the steps in `/etc/bash.bashrc.local`.

Advantages to this approach:
- It works.
- It creates a more minimal and predictable environment.

Disadvantages:
- Paths to modules.sh and modulefiles are hard-coded and need updates.
- The list of required modules may change over time.