(*)Revise offline tracer test parameters
  This commit updates the parameters, diag_table, and sample run script
fragments in ice_ocean_SIS2/Baltic_ALE_z_offline_tracers used to test the
offline tracer code, so that it once again runs without encountering errors,
when used along with a parallel commit to MOM6.  The altered files are not
used in the standard runs in the MOM6-examples test suite, and all reference
solutions that were working previously give bitwise identical answers.