Remove global_ALE/SLight configuration
  Removed the global_ALE/SLight configuration from MOM6_examples.  This was a
failed experiment attempting to deal with the high latitude haloclines with an
otherwise isopycnal coordinate, but it did not work very well after the first
winter, and it has never been used in practice.  Eliminating this unused
experiment will permit some of the related code within MOM6 to also be
eliminated.