C5: Update mkmf; libcoupler removed from ice-ocean
The link to the mkmf submodule was updated to a branch which includes
the C5 templates.

Note that the nvhpc template was introduced for the Nvidia compilers,
although for practical reasons the "pgi" name has been preserved.  Once
everything is working, we can flip the name to nvidia or nvhpc or
whatever we like.

Makefile.build was updated so that the ice-ocean MOM6 was build directly
with the coupler *.o files, rather than an external libcoupler.  This
was only necessary due to a quirk in the new `ftn` wrapper on C5, which
only provides the necessary flags when a file is provided.