*Set RHO_0 = 1024.0 in ocean_only/nonBous_global
  Change the setting of RHO_0 to 1024.0 in the nonBous_global test case so that
a number of mathematically equivalent revisions to the MOM6 code that are
implementing a fully non-Boussinesq option in MOM6 will not be flagged as
failing the automated regression tests.  This is the only non-Boussinesq test
case, but it is not being used actively for any scientific studies, so although
this commit changes answers, it should not impact any ongoing projects.