Swap DEFAULT_ANSWER_DATE for DEFAULT_2018_ANSWERS
  Replace lines setting DEFAULT_2018_ANSWERS in the MOM_input files for 12
experiments with equivalent values of DEFAULT_ANSWER_DATE.  This is a necessary
prelimiary to obsoleting all the _2018_ANSWERS runtime parameters.  Because of
the atypical way that the default for PORBAR_ANSWER_DATE is set, this also has
to be set explicitly to keep the previous settings.  All answers are bitwise
identical, but there are differences in the contents of the MOM_parameter_doc
files for these cases.