Parameter update for previous commit
The Makefile updates from the previous commit included an update to
MOM6, but did not include updates to the parameters.

Under our current rules, this was an error.  However, the more important
rule is that we do not break the history of dev/gfdl MOM6-examples.  So
these are being committed separately.  Hopefully it will not happen
again.

For the sake of bookkeeping:

NOAA-GFDL/MOM6@e1ea758: Add the debugging routine test_L_open_concave