+Idealized_hurricane input parameter updates
  Revised some of the settings in the MOM_input file for SCM_idealized_hurricane
to reflect how this configuration works.  Also added commented out settings in
the MOM_override for this test case to reflect the settings that would be needed
to give mathematically equivalent answers using the more standard "ideal_hurr"
value WIND_CONFIG, rather than "SCM_ideal_hurr".  This configuration will be
revised once some pending changes to the user/Idealized_hurricane.F90 code
are in place.  The answers in the SCM_idealized_hurricane test case are bitwise
identical.

  Also modified the MOM_input file for the idealized_hurricane test case to
avoid using IDL_HURR_SCM_BR_BENCH = True, which recreates an obvious bug, and
adds an as-yet unused parameter that will allow this test case to work as
intended with the pending code changes to Idealized_hurricane.F90.  This
change does change answers in that case, but it does not impact any of our usual
regression testing.

  The MOM_parameter_doc files were also updated for both the idealized_hurricane
test cases.