Explicitly define external directory of SIS2
The imminent introduction of the icebergs stub to SIS2, which is not
used in our CI tests, can override the libicebergs.a symbols, causing
the latter to never be used.  This can change answers (among other
things).

The glob `external/*` mkmf pathing is replaced with
`external/Icepack_interfaces`, so that any future icebergs stub will be
ignored.